refactor(accounting-plugin): single-source server↔vue response/report types
Declare the domain + report shapes once in src/shared/types.ts (isomorphic,
browser-safe) and import them from both the server and the Vue surface, instead
of hand-mirroring them in server/types.ts, server/report.ts and vue/api.ts.
Same-kind fix as #2334: adding a field to an API response can no longer leave
the Vue-side type silently stale — there is one declaration.

- New src/shared/types.ts: Account, AccountType, BALANCE_SHEET_ACCOUNT_TYPES,
  JournalLine, JournalEntry, JournalEntryKind, BookSummary, AccountBalance,
  ReportPeriod + report shapes (BalanceSheet(Section), ProfitLoss, Ledger(Row)).
- server/types.ts keeps only the server-only persistence shapes
  (AccountingConfig, MonthSnapshot); server files import domain/report types
  from ../shared/types.js.
- vue/api.ts drops its duplicate declarations and re-exports the shared types so
  the View / components keep their single ../api import surface.
- Extract the shared Map→sorted AccountBalance[] tail into a pure
  sortedBalancesFromMap helper (report.ts), reused by snapshotCache.mergeBalances;
  unit-tested.
- http.ts asyncHandler: pin the "intentional copy, not shared" reason (a plugin
  importing the host util would be an uphill import; the copies also diverge).

No public API shape change, no behaviour change, no version bump.

Closes #2411
