+++ packages/plugins/accounting-plugin/src/server/http.ts
@@ -2,8 +2,14 @@
 // self-contained (no host util imports). `asyncHandler` turns an
 // uncaught throw inside an async handler into a logged 500 carrying
 // only the caller-supplied fallback message — never the raw error text
-// (which could leak internals). Mirrors the host's
-// server/utils/asyncHandler.ts, scoped to this package's logger.
+// (which could leak internals).
+//
+// Deliberately NOT shared with the host's server/utils/asyncHandler.ts:
+// a plugin importing host code is an uphill import, forbidden by the
+// package-dependency-direction rule. This copy is also scoped to the
+// package logger and diverges — it forwards to Express `next(err)` on
+// `headersSent` (the host copy doesn't). Do not "de-dupe" it by reaching
+// into the host; lift it into @mulmoclaude/core if it ever must be shared.
 
 import type { Request, Response, NextFunction } from "express";
 import { log } from "./context.js";
