+++ packages/plugins/accounting-plugin/src/server/io.ts
@@ -15,7 +15,8 @@ import path from "node:path";
 import { defaultWorkspaceRoot } from "./context.js";
 import { ACCOUNTING_DIRS as WORKSPACE_DIRS, resolveFiscalYearEnd } from "../shared";
 import { writeJsonAtomic, isEnoent } from "./atomic.js";
-import type { AccountingConfig, Account, BookSummary, JournalEntry, MonthSnapshot } from "./types.js";
+import type { AccountingConfig, MonthSnapshot } from "./types.js";
+import type { Account, BookSummary, JournalEntry } from "../shared/types.js";
 
 const root = (workspaceRoot?: string): string => workspaceRoot ?? defaultWorkspaceRoot();
