+++ packages/plugins/accounting-plugin/test/accounting/test_accountNormalize.ts
@@ -7,7 +7,7 @@
 import { describe, it } from "node:test";
 import assert from "node:assert/strict";
 import { normalizeStoredAccount } from "../../src/server/accountNormalize.ts";
-import type { Account } from "../../src/server/types.ts";
+import type { Account } from "../../src/shared/types.ts";
 
 const BASE: Account = { code: "1500", name: "Equipment", type: "asset" };
