+++ packages/plugins/accounting-plugin/test/accounting/test_journal.ts
@@ -10,7 +10,7 @@ import {
   validateEntry,
   voidedIdSet,
 } from "../../src/server/journal.js";
-import type { Account, JournalEntry, JournalLine } from "../../src/server/types.js";
+import type { Account, JournalEntry, JournalLine } from "../../src/shared/types.js";
 
 const ACCOUNTS: Account[] = [
   { code: "1000", name: "Cash", type: "asset" },
